portfolio-backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.3
django-cors-headers==4.2.0
django-taggit==4.0.0
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.1
Pillow==10.0.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
whitenoise==6.5.0
django-seo==0.3.5
django-seo2==1.0.1